Deduplicate InputFile class in tests.
